1.0.3
Fixed more issues related to starting the IE driver. This time it should work...

Fixed xpath axis issues for some of the methods in GetElement and GetElements

Refactored some of the default options (if no arguments are passed) for the various drivers. They all use direct proxy now, which can for some drivers provide a performance boost